+++ pom.xml
@@ -47,7 +47,7 @@
             <dependency>
                 <groupId>com.forgerock.securebanking.uk</groupId>
                 <artifactId>securebanking-openbanking-uk-bom</artifactId>
-                <version>1.0.3-SNAPSHOT</version>
+                <version>1.1.3-SNAPSHOT</version>
                 <type>pom</type>
                 <scope>import</scope>
             </dependency>
@@ -67,12 +67,12 @@
         <repository>
             <id>maven.forgerock.org-community</id>
             <name>maven.forgerock.org-releases</name>
-            <url>https://maven.forgerock.org:443/repo/community</url>
+            <url>https://maven.forgerock.org/artifactory/community</url>
         </repository>
         <snapshotRepository>
             <id>maven.forgerock.org-community-snapshots</id>
             <name>maven.forgerock.org-snapshots</name>
-            <url>https://maven.forgerock.org:443/repo/community</url>
+            <url>https://maven.forgerock.org/artifactory/community</url>
         </snapshotRepository>
     </distributionManagement>
 
@@ -80,12 +80,12 @@
         <repository>
             <id>maven.forgerock.org-community</id>
             <name>maven.forgerock.org-releases</name>
-            <url>https://maven.forgerock.org:443/repo/community</url>
+            <url>https://maven.forgerock.org/artifactory/community</url>
         </repository>
         <repository>
             <id>maven.forgerock.org-community-snapshots</id>
             <name>maven.forgerock.org-snapshots</name>
-            <url>https://maven.forgerock.org:443/repo/community</url>
+            <url>https://maven.forgerock.org/artifactory/community</url>
         </repository>
         <repository>
             <id>central</id>
